git-dcloudfunction/requirements.txt
  # An example requirements file. If your function has other dependencies,
  # add them below
  flask
  python-dotenv
  functions-framework==3.*
  requests